Add files via upload
Vitaprint_generator.py generates the g-code for a simple scaffold and can be used as a testing tool for Vitaprint extruder. As a prerequisite you have to have Python 3 installed on your device in order to be able to run this script. Please read carefully the "USING THIS GENERATOR" section in the script header before attempting to run the script.